Replace data types of `acc_id`, `ascc_id`, `bcc_id`, `asccp_id`, `bccp_id`, `dt_id`, `code_list_id`, and `agency_id_list_id` with UUID char(36).

+++ score-http/src/main/java/org/oagi/score/repo/component/acc/AccReadRepository.java
@@ -137,7 +137,7 @@ public class AccReadRepository {
                     ccList.setLastUpdateUser(e.get(appUserUpdater.LOGIN_ID));
                     ccList.setLastUpdateTimestamp(Date.from(e.getValue("last_update_timestamp", LocalDateTime.class)
                             .atZone(ZoneId.systemDefault()).toInstant()));
-                    ccList.setId(e.get(ACC.ACC_ID).toBigInteger());
+                    ccList.setId(e.get(ACC.ACC_ID));
                     return  ccList;
                 });
     }
